refactor(DTFS2-6750): sign in link file names
