+++ portal-api/src/v1/users/sign-in-link.controller.test.js
@@ -1,10 +1,10 @@
 const { when } = require('jest-when');
-const { createAndSendAuthenticationToken } = require('./authentication-token.controller');
-const service = require('./authentication-token.service');
+const { createAndEmailSignInLink } = require('./sign-in-link.controller');
+const service = require('./sign-in-link.service');
 
-jest.mock('./authentication-token.service');
+jest.mock('./sign-in-link.service');
 
-describe('authentication token controller', () => {
+describe('sign in link controller', () => {
   const user = {
     _id: 'an id',
     email: 'an email',
@@ -26,39 +26,39 @@ describe('authentication token controller', () => {
     res.status.mockReturnThis();
   });
 
-  describe('createAndSendAuthenticationToken', () => {
+  describe('createAndEmailSignInLink', () => {
     it('should create and send a sign in code for req.user', async () => {
-      await createAndSendAuthenticationToken(req, res);
-      expect(service.createAndSendAuthenticationToken).toHaveBeenCalledWith(user);
+      await createAndEmailSignInLink(req, res);
+      expect(service.createAndEmailSignInLink).toHaveBeenCalledWith(user);
     });
 
     it('should respond with a 201 if the sign in code is sent', async () => {
-      await createAndSendAuthenticationToken(req, res);
+      await createAndEmailSignInLink(req, res);
       expect(res.status).toHaveBeenCalledWith(201);
     });
 
     it('should respond with an empty body if the sign in code is sent', async () => {
-      await createAndSendAuthenticationToken(req, res);
+      await createAndEmailSignInLink(req, res);
       expect(res.send).toHaveBeenCalledWith();
     });
 
     it('should respond with a 500 if the sign in code fails', async () => {
-      when(service.createAndSendAuthenticationToken)
+      when(service.createAndEmailSignInLink)
         .calledWith(user)
         .mockRejectedValueOnce(new Error());
 
-      await createAndSendAuthenticationToken(req, res);
+      await createAndEmailSignInLink(req, res);
 
       expect(res.status).toHaveBeenCalledWith(500);
     });
 
     it('should respond with the error message as a response body if the sign in code fails', async () => {
       const errorMessage = 'a test error';
-      when(service.createAndSendAuthenticationToken)
+      when(service.createAndEmailSignInLink)
         .calledWith(user)
         .mockRejectedValueOnce(new Error(errorMessage));
 
-      await createAndSendAuthenticationToken(req, res);
+      await createAndEmailSignInLink(req, res);
 
       expect(res.send).toHaveBeenCalledWith({
         error: 'Internal Server Error',
