+++ portal-api/src/v1/users/sign-in-link.service.js
@@ -4,7 +4,7 @@ const sendEmail = require('../email');
 const { EMAIL_TEMPLATE_IDS, SIGN_IN_LINK_EXPIRY_MINUTES } = require('../../constants');
 const db = require('../../drivers/db-client');
 
-module.exports.createAndSendAuthenticationToken = async (user) => {
+module.exports.createAndEmailSignInLink = async (user) => {
   const {
     _id: userId,
     email: userEmail,
@@ -24,8 +24,7 @@ module.exports.createAndSendAuthenticationToken = async (user) => {
 
 async function createSignInCode() {
   try {
-    const signInCode = crypto.randomBytes(32).toString('hex');
-    return signInCode;
+    return crypto.randomBytes(32).toString('hex');
   } catch (e) {
     const error = new Error('Failed to create a sign in code.');
     error.cause = e;
@@ -39,7 +38,6 @@ async function saveSignInCodeHashAndSalt({ userId, signInCode }) {
     // TODO DTFS2-6750: check best practice for pbkdf2
     const hash = crypto.pbkdf2Sync(signInCode, salt, 10000, 64, 'sha512').toString('hex');
 
-    // TODO DTFS2-6750: need to unit test error handling for db call (could not do in api test)
     // TODO DTFS2-6750: db layer
     const userCollection = await db.getCollection('users');
     await userCollection.updateOne({ _id: { $eq: ObjectId(userId) } }, { $set: { signInCode: { hash, salt } } });
@@ -63,7 +61,7 @@ async function sendSignInLinkEmail({ userEmail, userFirstName, userLastName, sig
       },
     );
   } catch (e) {
-    const error = new Error('Failed to send the sign in code via email.');
+    const error = new Error('Failed to email the sign in code.');
     error.cause = e;
     throw error;
   }
